+++ src/main/java/com/szq/hotel/entity/bo/Order.java
@@ -26,7 +26,7 @@ public class Order  extends BaseModel {
     private Integer membersId;//会员id
     private Integer memberIdOrOrganizationId;//会员id 合作机构id
     private String subscribeRemark;//预约备注
-    private double totalPrice;//总房价
+    private BigDecimal totalPrice;//总房价
     private Date createTime;
     private Date updateTime;
 
@@ -166,11 +166,11 @@ public class Order  extends BaseModel {
         this.subscribeRemark = subscribeRemark;
     }
 
-    public double getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(double totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
